Updatedd
s

+++ data_preprocessing/sent_to_tuples.py
@@ -2,14 +2,14 @@
 def sent_to_tuple(sent:list,wind_size = 3,left_bias = True):
     output_list = []
     if wind_size % 2 == 1:
-        left = (wind_size - 1)/2
+        left = int((wind_size - 1)/2)
         right = left
     else:
         if left_bias:
-            left = wind_size/2
+            left = int(wind_size/2)
             right = left -1
         else:
-            right = wind_size/2
+            right = int(wind_size/2)
             left = right -1
 
     for i in range(len(sent)):
